blog-flask-api/requirement.txt
alembic==1.11.1
Babel==2.12.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
dominate==2.8.0
ecdsa==0.18.0
elastic-transport==8.4.0
elasticsearch==8.8.2
email-validator==2.0.0.post2
Flask==2.3.2
flask-babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
pyparsing==3.1.0
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
